tweaked players with voiceConnection check, added Utility class for general utilty methods, added pretty loader to search command and paginated queue viewing

+++ commands/music/moe-radio.js
@@ -41,16 +41,20 @@ module.exports = class MoeRadioCommand extends Command {
     {
         let playingMessage = null;
         this.client.moe_radio.on('streaming', async (embed, guild) => {
-            if (playingMessage && playingMessage.deletable) playingMessage.delete();
-            let channel = playingMessage ? playingMessage.channel : guild.channels.find('type', 'text');
-            if (channel && embed !== null) playingMessage = (await channel.send('', {embed: embed}));
-            else console.log(`No text channel found for guild ${guild.id}/${guild.name} to display radio playing embed.`)
+            if (guild.voiceConnection) {
+                if (playingMessage && playingMessage.deletable) playingMessage.delete();
+                let channel = playingMessage ? playingMessage.channel : guild.channels.find('type', 'text');
+                if (channel && embed !== null) playingMessage = (await channel.send('', {embed: embed}));
+                else console.log(`No text channel found for guild ${guild.id}/${guild.name} to display radio playing embed.`)
+            }
         });
 
         this.client.moe_radio.on('stream', (text, guild) => {
-            let channel = guild.channels.find('type', 'text');
-            if (channel && text) channel.send(text);
-            else console.log(`No text channel found for guild ${guild.id}/${guild.name} to display music playing embed.`)
+            if (guild.voiceConnection) {
+                let channel = guild.channels.find('type', 'text');
+                if (channel && text) channel.send(text);
+                else console.log(`No text channel found for guild ${guild.id}/${guild.name} to display music playing embed.`)
+            }
         });
 
         this.client.moe_radio.on('error', text => { throw text; });
